@@ -1,14 +1,14 @@
 name: flutter_date_range_picker
 description: Flutter package for selecting date ranges with input, dialog, and calendar options.
-version: 0.0.14
-homepage: https://github.com/GaspardMerten/date_range_picker
+version: 0.0.15
+homepage: https://github.com/zonaro/date_range_picker
 
 environment:
-  sdk: '>=2.19.2 <3.0.0'
-  flutter: ">=1.17.0"
+  sdk: '>=3.3.0 <4.0.0'
+ 
 
 dependencies:
-  intl: ^0.18.0
+  intl: 
   flutter:
     sdk: flutter
 
