@@ -1,5 +1,5 @@
+import 'package:flutter/widgets.dart';
 import 'package:intl/intl.dart';
 
 /// Returns the default week days as strings (using intl).
-List<String> defaultWeekDays() =>
-    DateFormat.E().dateSymbols.WEEKDAYS.map((e) => e.substring(0, 3)).toList();
+List<String> defaultWeekDays([Locale? locale]) => DateFormat.E(locale?.languageCode).dateSymbols.WEEKDAYS.map((e) => e.substring(0, 3)).toList();
