@@ -4,11 +4,13 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('DayNamesRow Widget', () {
-    testWidgets('renders correctly with default week days',
-        (WidgetTester tester) async {
+    testWidgets('renders correctly with default week days', (WidgetTester tester) async {
       await tester.pumpWidget(MaterialApp(
         home: Scaffold(
-          body: DayNamesRow(textStyle: kTheme.dayNameTextStyle),
+          body: DayNamesRow(
+            textStyle: kTheme.dayNameTextStyle,
+            weekDays: const [],
+          ),
         ),
       ));
 
@@ -19,10 +21,8 @@ void main() {
   });
 
   group('DateRangePickerWidget', () {
-    testWidgets('renders correctly and initializes with initial dateRange',
-        (WidgetTester tester) async {
-      final initialDateRange =
-          DateRange(DateTime(2023, 1, 1), DateTime(2023, 1, 5));
+    testWidgets('renders correctly and initializes with initial dateRange', (WidgetTester tester) async {
+      final initialDateRange = DateRange(DateTime(2023, 1, 1), DateTime(2023, 1, 5));
       final minDate = DateTime(2022, 1, 1);
       final maxDate = DateTime(2023, 12, 31);
 
